backend/requirements.txt
opencv-python
fonttools
python-multipart
fastapi
uvicorn
python-dotenv
requests
pydantic
sqlalchemy
psycopg2-binary
redis
boto3
minio
openai
serpapi
pymupdf
Pillow
numpy
pandas
structlog
pydantic_settings
pdf2docx>=0.5.8
supabase>=2.0.0
aiofiles
python-jose
httpx
python-docx